TELERIK, withCredentials, LOGOUT, IS4 cors policies

+++ Services/HttpServices/GenericHttpService.cs
@@ -21,6 +21,8 @@ using Microsoft.IdentityModel.Protocols.OpenIdConnect;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Authentication.Cookies;
 using IdentityModel.Client;
+using Microsoft.AspNetCore.Authentication.OpenIdConnect;
+using Microsoft.AspNetCore.Http.Extensions;
 
 namespace Services.HttpServices
 {
@@ -41,8 +43,8 @@ namespace Services.HttpServices
         Task<(HttpStatusCode sc, T entidad)> HttPostAsync(T entidadDto);
         Task<(HttpStatusCode sc, T entidad)> HttpPutAsync(T entidadDto, string id);
         Task<HttpStatusCode> HttpDeleteAsync(string id);
-        Task<HttpClient> GetHttpClientWithToken();
-        Task<(string at, int error)> RefreshToKen();
+        Task<(HttpClient, HttpStatusCode)> GetHttpClientWithToken();
+        Task<(string at, HttpStatusCode sc)> RefreshToKen();
 
         public struct CustomResponse
         {
@@ -64,28 +66,33 @@ namespace Services.HttpServices
         }
 
     }
-    public class GenericHttpService<T> :  IGenericHttpService<T> where T : class
+    public class GenericHttpService<T> : IGenericHttpService<T> where T : class
     {
         private readonly IHttpClientFactory _httpClientFactory;
         private readonly IHttpContextAccessor _httpContextAccessor;
         private readonly TokenClient _tokenClient;
+   
 
         public GenericHttpService(IHttpClientFactory httpClientFactory, IHttpContextAccessor httpContextAccessor, TokenClient tokenClient)
         {
             _httpClientFactory = httpClientFactory;
             _httpContextAccessor = httpContextAccessor;
             _tokenClient = tokenClient;
+           
+
+       
         }
        
 
         public string controlador {private get; set; }
         public string cliente {private get; set; }
 
-
        
 
-        public async Task<(string at, int error)> RefreshToKen()
+        public async Task<(string at, HttpStatusCode sc)> RefreshToKen()
         {
+
+            var fake = "kk";
            
             var rt = await _httpContextAccessor.HttpContext.GetTokenAsync(OpenIdConnectParameterNames.RefreshToken);
             var tokenResult = await _tokenClient.RequestRefreshTokenAsync(rt);
@@ -107,35 +114,68 @@ namespace Services.HttpServices
                 //properties.AllowRefresh = true;
                 //ShouldRenew = true;
 
-                await authService.SignInAsync(_httpContextAccessor.HttpContext, CookieAuthenticationDefaults.AuthenticationScheme, authenticateResult.Principal, authenticateResult.Properties);             
+                await authService.SignInAsync(_httpContextAccessor.HttpContext, CookieAuthenticationDefaults.AuthenticationScheme, authenticateResult.Principal, authenticateResult.Properties);
 
-                return (await _httpContextAccessor.HttpContext.GetTokenAsync("access_token"), 0);
+                return (await _httpContextAccessor.HttpContext.GetTokenAsync("access_token"), tokenResult.HttpStatusCode); // HttpStatusCode.OK);
             }
 
-            return  (await _httpContextAccessor.HttpContext.GetTokenAsync("access_token"), 1);
+
+
+            ///await _httpContextAccessor.HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
+            //await _httpContextAccessor.HttpContext.SignOutAsync(OpenIdConnectDefaults.AuthenticationScheme, new AuthenticationProperties() { RedirectUri = "home/prueba" });
+            /////
+            ////_httpContextAccessor.HttpContext.Abort();
+            ////await _httpContextAccessor.HttpContext.SignOutAsync(CookieAuthenticationDefaults.AuthenticationScheme);
+            ////new SignOutResult(OpenIdConnectDefaults.AuthenticationScheme, new AuthenticationProperties { RedirectUri = "Home/login" });
+            //var rq = _httpContextAccessor.HttpContext.Request;
+            //var url = UriHelper.BuildAbsolute(rq.Scheme, rq.Host, rq.PathBase, "/home/login");
+            //_httpContextAccessor.HttpContext.Response.Redirect(UriHelper.BuildAbsolute(rq.Scheme, rq.Host, rq.PathBase, "/home/about"));
+
+            var authServiceOut = _httpContextAccessor.HttpContext.RequestServices.GetService<IAuthenticationService>();
+            await authServiceOut.SignOutAsync(_httpContextAccessor.HttpContext, CookieAuthenticationDefaults.AuthenticationScheme, new AuthenticationProperties() { });
+            await authServiceOut.SignOutAsync(_httpContextAccessor.HttpContext, OpenIdConnectDefaults.AuthenticationScheme, new AuthenticationProperties() { RedirectUri = "home/login" });
+
+            // await _httpContextAccessor.HttpContext.ChallengeAsync(OpenIdConnectDefaults.AuthenticationScheme, new AuthenticationProperties() { RedirectUri = "home/about" });
+
+            //uriHelper.RouteUrl("", new { controller = "Home", action = "Login" }) });
+
+            //var authServiceOut = _httpContextAccessor.HttpContext.RequestServices.GetService<IAuthenticationService>();
+            //await authServiceOut.SignOutAsync(_httpContextAccessor.HttpContext, CookieAuthenticationDefaults.AuthenticationScheme, new AuthenticationProperties { RedirectUri = "home/about" });
+            //await _httpContextAccessor.HttpContext.ChallengeAsync()
+
+            return (string.Empty, (HttpStatusCode)_httpContextAccessor.HttpContext.Response.StatusCode);
+
+            //return (await _httpContextAccessor.HttpContext.GetTokenAsync("access_token"), 1);
         }
 
-        public async Task<HttpClient> GetHttpClientWithToken()
+        public async Task<(HttpClient, HttpStatusCode)> GetHttpClientWithToken()
         {
-            var accessToken = await _httpContextAccessor.HttpContext.GetTokenAsync(OpenIdConnectParameterNames.AccessToken);
-            int error;
+            var (accessToken, sc) = (await _httpContextAccessor.HttpContext.GetTokenAsync(OpenIdConnectParameterNames.AccessToken), HttpStatusCode.OK);
+            //int err = 0;
 
             DateTime accessTokenExpiresAt = DateTime.Parse(await _httpContextAccessor.HttpContext.GetTokenAsync("expires_at"), CultureInfo.InvariantCulture, DateTimeStyles.RoundtripKind);
-            if(DateTime.Now.CompareTo(accessTokenExpiresAt.AddMinutes(-1)) >= 0) { (accessToken, error) = await RefreshToKen(); }
+            if(DateTime.Now.CompareTo(accessTokenExpiresAt.AddMinutes(-1)) >= 0) {(accessToken,  sc) = await RefreshToKen(); }
 
-            var client = _httpClientFactory.CreateClient($"{cliente}");
-            client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(IdentityServerAuthenticationDefaults.AuthenticationScheme , accessToken);
+            if (sc == HttpStatusCode.Redirect) return (null, sc);
 
-            return client;
+            //if (err == 0)
+            //{
+                var client = _httpClientFactory.CreateClient($"{cliente}");
+                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(IdentityServerAuthenticationDefaults.AuthenticationScheme, accessToken);
+
+                return (client, HttpStatusCode.OK);
+            //}
+            //return(null, HttpStatusCode.Unauthorized);
            
         }
         
         public async Task<(HttpStatusCode sc, gridDto<T> gridDto)> HttpGetAsync(QueryString query)
         {
           
-            var client = await GetHttpClientWithToken();
+            var(client, sc) = await GetHttpClientWithToken();
+            if (sc == HttpStatusCode.Redirect) return (sc, null);
 
-            var result = await client.GetAsync($"{controlador}{query}"); //.Result.Content.ReadAsStringAsync();
+            var result = await client.GetAsync($"{controlador}{query}"); 
 
             if (result.StatusCode == HttpStatusCode.OK)
             {
@@ -159,8 +199,9 @@ namespace Services.HttpServices
 
         public async Task<(HttpStatusCode sc, gridDto<T> gridDto)> HttpGetDetailAsync(QueryString query, string id)
         {
-           
-            var client = await GetHttpClientWithToken();
+
+            var (client, sc) = await GetHttpClientWithToken();
+            if (sc == HttpStatusCode.Redirect) return (sc, null);
 
             var result = await client.GetAsync($"{controlador}/{id}{query}"); 
 
@@ -176,10 +217,11 @@ namespace Services.HttpServices
         public async Task<KeyValuePair<HttpStatusCode, List<KeyValuePair<string, object>>>> HttpCbosAsync()
         {
 
-            var client = await GetHttpClientWithToken();
+            var (client, sc) = await GetHttpClientWithToken();
+            if (sc == HttpStatusCode.Redirect) return new KeyValuePair<HttpStatusCode, List<KeyValuePair<string, object>>>(sc, null);
 
             HttpResponseMessage result = await client.GetAsync($"Combos/{controlador}Cbos");
-            if (result.StatusCode == System.Net.HttpStatusCode.OK)
+            if (result.StatusCode == HttpStatusCode.OK)
             {
                 var entidad = JsonConvert.DeserializeObject<List<KeyValuePair<string, object>>>(result.Content.ReadAsStringAsync().Result, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.All });
                
@@ -195,11 +237,12 @@ namespace Services.HttpServices
         public async Task<(HttpStatusCode sc, T entidad)> HttPostAsync(T entidadDto)
         {
             //var client = _httpClientFactory.CreateClient($"{cliente}");
-            var client = await GetHttpClientWithToken();
+            var (client, sc) = await GetHttpClientWithToken();
+            if (sc == HttpStatusCode.Redirect) return (sc, null);
             using (var content = new StringContent(JsonConvert.SerializeObject(entidadDto), Encoding.UTF8, "application/json"))
             {
-                HttpResponseMessage result = await client.PostAsync($"{controlador}", content); //controlador
-                if (result.StatusCode == System.Net.HttpStatusCode.Created)
+                HttpResponseMessage result = await client.PostAsync($"{controlador}", content); 
+                if (result.StatusCode == HttpStatusCode.Created)
                 {                  
                     var entidad = JsonConvert.DeserializeObject<T>(result.Content.ReadAsStringAsync().Result, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.All });
 
@@ -214,12 +257,13 @@ namespace Services.HttpServices
 
         public async Task<(HttpStatusCode sc, T entidad)> HttpPutAsync(T entidadDto, string id)
         {
-            
-            var client = await GetHttpClientWithToken();
+
+            var (client, sc) = await GetHttpClientWithToken();
+            if (sc == HttpStatusCode.Redirect) return (sc, null);
             using (var content = new StringContent(JsonConvert.SerializeObject(entidadDto), Encoding.UTF8, "application/json"))
             {
                 HttpResponseMessage result = await client.PutAsync($"{controlador}/{id}", content); 
-                if (result.StatusCode == System.Net.HttpStatusCode.Created)
+                if (result.StatusCode == HttpStatusCode.Created)
                 {
                     var entidad = JsonConvert.DeserializeObject<T>(result.Content.ReadAsStringAsync().Result, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.All });
                    
@@ -235,9 +279,10 @@ namespace Services.HttpServices
         public async Task<HttpStatusCode> HttpDeleteAsync(string id)
         {
             //var client = _httpClientFactory.CreateClient($"{cliente}");
-            var client = await GetHttpClientWithToken();
-            HttpResponseMessage result = await client.DeleteAsync($"{controlador}/{id}");
+            var (client, sc) = await GetHttpClientWithToken();
+            if (sc == HttpStatusCode.Redirect) return (sc);
 
+            HttpResponseMessage result = await client.DeleteAsync($"{controlador}/{id}");
             return (result.StatusCode);
            
         }
